backend/Lost/requirements.txt
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==2.1
django-extensions==2.2.9
django-filter==2.2.0
django-sslserver==0.22
django-werkzeug-debugger-runserver==0.3.1
djangorestframework==3.11.0
idna==2.9
Markdown==3.1.1
mysqlclient==1.4.4
Pillow==7.0.0
pycparser==2.20
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
